requirements.txt
asgiref==3.2.10
Brotli==1.1.0
CacheControl==0.14.0
cachetools==5.4.0
campay==1.0.9
certifi==2020.6.20
cffi==1.17.1
channels==4.1.0
chardet==3.0.4
charset-normalizer==3.3.2
cryptography==43.0.1
cssselect2==0.7.0
defusedxml==0.6.0
Django==3.0.8
django-cors-headers==4.4.0
django-crispy-forms==1.9.1
django-debug-toolbar==2.2
django-environ==0.4.5
django-model-utils==4.0.0
django-redis==4.12.1
django-rest-framework==0.1.0
djangorestframework==3.11.0
djangorestframework-simplejwt==5.3.1
firebase-admin==6.5.0
fonttools==4.53.1
fpdf==1.7.2
google-api-core==2.19.1
google-api-python-client==2.138.0
google-auth==2.32.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
google-cloud-core==2.4.1
google-cloud-firestore==2.19.0
google-cloud-storage==2.18.2
google-crc32c==1.6.0
google-resumable-media==2.7.2
googleapis-common-protos==1.63.2
grpcio==1.66.2
grpcio-status==1.66.2
html5lib==1.1
httplib2==0.22.0
idna==2.10
msgpack==1.1.0
oauthlib==3.1.0
pillow==10.4.0
proto-plus==1.24.0
protobuf==5.27.2
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.20
pydyf==0.11.0
PyJWT==2.9.0
pyparsing==3.1.2
pyphen==0.16.0
python-decouple==3.3
python-slugify==4.0.1
python3-openid==3.2.0
pytz==2020.1
redis==5.2.0
reportlab==4.2.2
requests==2.24.0
requests-oauthlib==1.3.0
rsa==4.9
setuptools==75.6.0
six==1.15.0
sqlparse==0.3.1
text-unidecode==1.3
tinycss2==1.3.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==1.25.9
weasyprint==62.3
webencodings==0.5.1
zopfli==0.2.3
